+++ CHANGELOG.md
@@ -9,6 +9,7 @@
     - SPI *(tx only)*
     - USART *(tx only)*
     - SDIO *(STM32F4 devices only)*
+    - Basic interface for System Clock and PLL settings.
 - Adds chip drivers for:
     - Display Controller *ILI9341*
     - Display Controller *UC1611S*